==> R/07_scales_diff.r <==
library(pacman)

p_load(here, tidyverse, shapes, stringr, geomorph, scales, grid)

source(here::here('R', 'helper04_df2array.r'))
source(here::here('R', 'helper06_multiplot.r'))
source(here::here('R', 'helper09_compare_groups.r'))

theme_set(theme_minimal())

set.seed(420)

# create null distribution
rand_meandist <- function(shape, f) {
  tt <- sample(f)
  ss <- split.land(shape, factor(tt))
  sm <- map(ss, mshape)
  null <- procdist(sm[[1]], sm[[2]])
  null
}


# full clean data as generated by 01_supervised_mung.r
adult <- read_rds(here::here('data', 'turtle_plastron_data_clean.rds'))

# full landmark data after GPA as generated by 01_supervised_mung.r
fit <- read_rds(here::here('data', 'turtle_plastron_gpa.rds'))

# there are three dead columns that give problems with tibble-s
adult_scales <- 
  adult[, -((ncol(adult) - 2):ncol(adult))] %>%
  as_tibble(.) %>%
  mutate(spec = as.character(spec))


# combine adult with new scute information

# need to drop scute NAs from fit$rotated, which isn't obvious
trm <- is.na(adult_scales$scute)

adult_scales <- adult_scales %>% drop_na(scute)
scute_shape <- fit$rotated[, , !trm]

# adult_scales$inguinal_scute describes two groups
# those with scutes (1)
# those without scutes (0)
# is there significant difference in mean shape between these groups?
# null is from randomizing scute state, then comparing the RNG groups

# split by scute state
scute_split <- split.land(scute_shape, factor(adult_scales$scute))

# calculate mean scute state shape
# then calculate distance between mean scute state shapes
scute_mean <- map(scute_split, mshape)

scute_mean_dist <- procdist(scute_mean[[1]], scute_mean[[2]])


null <- rerun(1000, rand_meandist(scute_shape, factor(adult_scales$scute))) %>%
  tibble(null = .) %>%                 # rerun is type safe
  unnest()                             # need to unpack list column

dist_p <- 
  null %>%
  dplyr::summarize(sum(scute_mean_dist > null) / length(null) * 100) %>%
  pull()


scute_null_plot <- 
  null %>%
  ggplot(aes(x = null)) +
  geom_histogram(colour = 'darkgrey', fill = 'lightgrey') +
  geom_vline(xintercept = scute_mean_dist, size = 2) +
  scale_x_continuous(breaks = pretty_breaks(4)) +
  labs(x = 'Procrustes distance', y = 'Frequency',
       subtitle = paste0('Observed greater than ', dist_p, '% of null values'))
ggsave(filename = here::here('doc', 'figure', 'scales_test_hist.pdf'),
       plot = scute_null_plot,
       width = 4, height = 3)


# now do all the above but first break by group
gg <- c('sp10.1', 'sp10.2', 'sp10.3', 'sp14.1', 'sp14.2', 'morph')

# this form is mostly because i want this done now, not tomorrow
out <- list(p_value = c(), 
            graph = list())
for(ii in seq(length(gg))) {
  short <- 
    adult_scales %>%
    drop_na(gg[ii], scute)

  scute_split <- 
    split.land(scute_shape, factor(short$scute))

  scute_mean <- map(scute_split, mshape)
  scute_mean_dist <- procdist(scute_mean[[1]], scute_mean[[2]])

  null <- rerun(1000, rand_meandist(scute_shape, 
                                    factor(adult_scales$scute))) %>%
    tibble(null = .) %>%                 # rerun is type safe
    unnest()                             # need to unpack list column

  dist_p <- 
    null %>%
    dplyr::summarize(sum(scute_mean_dist > null) / length(null) * 100) %>%
    pull()

  scute_null_plot <- 
    null %>%
    ggplot(aes(x = null)) +
    geom_histogram(colour = 'darkgrey', fill = 'lightgrey') +
    geom_vline(xintercept = scute_mean_dist, size = 2) +
    scale_x_continuous(breaks = pretty_breaks(4)) +
    labs(x = 'Procrustes distance', y = 'Frequency',
         subtitle = paste0(gg[ii], 
                           ': Obs greater than ', 
                           dist_p, 
                           '% of null'))

  out$p_value[ii] <- dist_p
  out$graph[[ii]] <- scute_null_plot
}


walk2(out$graph, gg,
     ~ ggsave(filename = paste0(here::here('doc', 'figure'), 
                                '/scales_test_hist_', .y, '.pdf'),
              plot = .x,
              width = 10, height = 8))
